add license headers, move add_spdx_header.py and use pathlib instead of os for compatability
