+++ tt-metal-llama3-70b/src/scripts/jwt_util.py
@@ -1,3 +1,7 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
 #!/usr/bin/env python3
 import argparse
 import json
