+++ tt-metal-llama3-70b/src/test_inference_api_alpaca_eval.py
@@ -1,3 +1,7 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
 import os
 import getpass
 import threading
