+++ tt-metal-mistral-7b/src/mistral_7b_alpaca_test.py
@@ -1,3 +1,7 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
 import os
 from collections import defaultdict
