+++ tt-metal-mistral-7b/src/test_mistral_7b_backend.py
@@ -1,3 +1,7 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
 import queue
 import os
 from pathlib import Path
